Adding Admin Functionalities

+++ src/DeleteBook.java
@@ -1,18 +1,55 @@
-import java.util.Scanner;
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
 public class DeleteBook implements IOOperation{
     @Override
     public void operation(Database database, User user) {
-        Scanner scanner = new Scanner(System.in);
-        System.out.println("\nEnter book title: ");
-        String booktitle = scanner.next();
-        int index = database.getBook(booktitle);
-        if (index>-1){
-            database.deleteBook(index);
-            System.out.println("Book deleted successfully!");
-        } else{
-            System.out.println("Book doesn't exist!");
-        }
-        user.menu(database,user);
+
+        JFrame frame = Main.frame(400,210);
+        frame.setLayout(new BorderLayout());
+
+        JLabel title = Main.title("Delete Book");
+        frame.getContentPane().add(title,BorderLayout.NORTH);
+
+        JPanel panel = new JPanel(new GridLayout(2,2,15,15));
+        panel.setBorder(BorderFactory.createEmptyBorder(0,20,20,20));
+
+        JLabel label = Main.label("Book Title:");
+        JTextField book_title = Main.textField();
+        JButton delete = Main.button("Delete Book");
+        JButton cancel = Main.button("Cancel");
+        panel.add(label);
+        panel.add(book_title);
+        panel.add(delete);
+        panel.add(cancel);
+
+        delete.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                if (book_title.getText().toString().matches("")){
+                    JOptionPane.showMessageDialog(new JFrame(),"Book Title " +
+                            "should be provided!");
+                    return;
+                }
+                int index = database.getBook(book_title.getText().toString());
+                if (index>-1){
+                    database.deleteBook(index);
+                    JOptionPane.showMessageDialog(new JFrame(),"Book deleted successfully!");
+                    frame.dispose();
+                } else{
+                    JOptionPane.showMessageDialog(new JFrame(),"Book doesn't exist!");
+                }
+            }
+        });
+        cancel.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                frame.dispose();
+            }
+        });
+
+        frame.getContentPane().add(panel,BorderLayout.CENTER);
     }
 }
